Create a version util

Remove unused ipc event
